Fase 1: corrige dos bugs de contenido encontrados con arnes de generacion
Se generaron 9900 ejercicios (15 habilidades x formatos x 300) verificando
que la respuesta correcta este entre las opciones, que haya 4 opciones y que
no se repitan. Aparecieron dos fallas reales:

1. moda: en el caso "no hay moda" los distractores colapsaban (con 2,4,6,8,10
   la media y la mediana son ambas 6) y el ejercicio quedaba con 3 opciones.
   Ahora se acumulan mas candidatos y se rellena de forma determinista.

2. angulos en formato INVERSO: el distractor "repetir el angulo dado" usaba el
   angulo a calcular en vez del que ve el alumno, asi que coincidia con la
   respuesta correcta; el relleno automatico producia valores artificiales
   como 121 grados. Ahora los distractores se construyen sobre el dato visible.
   Se excluyen 45 y 90 grados del suplementario por la misma razon.

Tambien se cambia el caso numerico de graficas a numOpts, que garantiza 4
opciones distintas.

+++ android/app/src/main/java/com/controlparental/kioscosuave/curriculum/Sec1Porcentajes.kt
@@ -53,9 +53,9 @@ internal object Sec1Porcentajes {
             question = question,
             options = Gen.numOpts(
                 ans,
-                base / dec,            // error: dividir en vez de multiplicar
-                base * (1 - dec),      // error: calcular el complemento
-                base - p               // error: restar el número del porcentaje
+                base / dec,               // error: dividir en vez de multiplicar
+                base * (1 - dec),         // error: calcular el complemento
+                (base - p).toDouble()     // error: restar el número del porcentaje
             ),
             answer = Gen.fmt(ans),
             steps = listOf(
@@ -101,7 +101,7 @@ internal object Sec1Porcentajes {
         return MathQuestion(
             question = question,
             options = Gen.numOpts(
-                total,
+                total.toDouble(),
                 part * p / 100.0,      // error: volver a aplicar el %
                 part * (100 - p) / 100.0,
                 part + p               // error: sumar el % al dato
